aurel/expose-vector-id: Move VectorId to common
