+++ iris-mpc-common/src/lib.rs
@@ -10,6 +10,7 @@ pub mod job;
 pub mod shamir;
 #[cfg(feature = "helpers")]
 pub mod test;
+pub mod vector_id;
 
 pub const IRIS_CODE_LENGTH: usize = 12_800;
 pub const MASK_CODE_LENGTH: usize = 6_400;
